braquiterapia_app/requirements.txt
Flask==3.0.0
openpyxl==3.1.2
PyMuPDF==1.23.8
Pillow==10.1.0
python-dateutil==2.8.2
